fix(web): enhance predicted winner logic for toss-up races and update UI accordingly
