+++ services/races-api/schemas.py
@@ -1,5 +1,3 @@
-from typing import Dict, List
-
 from pydantic import BaseModel, Field
 
 
@@ -22,19 +20,19 @@ class RaceForecastSummary(BaseModel):
     predicted_winner_name: str | None = None
     predicted_winner_party: str | None = None
     win_probability: float | None = None
-    party_probabilities: Dict[str, float] = Field(default_factory=dict)
+    party_probabilities: dict[str, float] = Field(default_factory=dict)
     margin_estimate: float | None = None
     rating: str | None = None
     confidence: str | None = None
     rationale: str | None = None
     takeaway: str | None = None
-    key_reasons: List[str] = Field(default_factory=list)
+    key_reasons: list[str] = Field(default_factory=list)
     uncertainty: str | None = None
     based_on_poll_count: int = 0
     generated_at: str | None = None
     model: str | None = None
-    source_urls: List[str] = Field(default_factory=list)
-    market_signals: List[Dict] = Field(default_factory=list)
+    source_urls: list[str] = Field(default_factory=list)
+    market_signals: list[dict] = Field(default_factory=list)
 
 
 class RaceSummary(BaseModel):
