+++ services/races-api/simple_publish_service.py
@@ -10,7 +10,6 @@ import os
 import threading
 import time
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
 
 from shared.models import RaceJSON
 
@@ -25,7 +24,7 @@ def _is_published_race_blob(blob_name: str) -> bool:
     return blob_name.startswith("races/") and blob_name.endswith(".json") and blob_name != _SUMMARIES_BLOB
 
 
-def _normalize_summary_index(payload) -> Optional[List[Dict]]:
+def _normalize_summary_index(payload) -> list[dict] | None:
     if not isinstance(payload, list):
         return None
     return [item for item in payload if isinstance(item, dict) and item.get("id")]
